requirement.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic[email]
alembic==1.13.2
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
motor==3.4.0
pymongo==4.6.0
elasticsearch==8.11.1
aiohttp==3.13.2
